refactor(frontend): support Long type for item prices

+++ Backend/src/main/java/com/secondhand/backend/dto/FavoriteResponse.java
@@ -4,13 +4,13 @@ public class FavoriteResponse {
     private Long id;
     private Long itemId;
     private String itemTitle;
-    private double itemPrice;
+    private Long itemPrice;
     private String itemStatus;
     private Long userId;
 
     public FavoriteResponse() {}
 
-    public FavoriteResponse(Long id, Long itemId, String itemTitle, double itemPrice, String itemStatus, Long userId) {
+    public FavoriteResponse(Long id, Long itemId, String itemTitle, Long itemPrice, String itemStatus, Long userId) {
         this.id = id;
         this.itemId = itemId;
         this.itemTitle = itemTitle;
@@ -22,14 +22,14 @@ public class FavoriteResponse {
     public Long getId() { return id; }
     public Long getItemId() { return itemId; }
     public String getItemTitle() { return itemTitle; }
-    public double getItemPrice() { return itemPrice; }
+    public Long getItemPrice() { return itemPrice; }
     public String getItemStatus() { return itemStatus; }
     public Long getUserId() { return userId; }
 
     public void setId(Long id) { this.id = id; }
     public void setItemId(Long itemId) { this.itemId = itemId; }
     public void setItemTitle(String itemTitle) { this.itemTitle = itemTitle; }
-    public void setItemPrice(double itemPrice) { this.itemPrice = itemPrice; }
+    public void setItemPrice(Long itemPrice) { this.itemPrice = itemPrice; }
     public void setItemStatus(String itemStatus) { this.itemStatus = itemStatus; }
     public void setUserId(Long userId) { this.userId = userId; }
 }
